% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fcts.R
\name{afedR_get_data_file}
\alias{afedR_get_data_file}
\title{Get path to data file}
\usage{
afedR_get_data_file(name_dataset)
}
\arguments{
\item{name_dataset}{Name of the dataset filename (see \link{afedR_list_available_data} for more details)}
}
\value{
A dataframe with the data
}
\description{
This is a helper function of book "Analyzing Financial and Economic Data with R" by Marcelo S. Perlin.
With this function you'll be able to read the tables used in the book. Every table is located by its names.
In order to fin available tables, use function \link{afedR_list_available_data}.
}
\examples{
df <- afedR_get_data_file('grunfeld.csv')
}
