% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R
\name{afedR_unzip_book_file}
\alias{afedR_unzip_book_file}
\title{Donwload and unzip book file}
\usage{
afedR_unzip_book_file(path_to_unzip = "~")
}
\arguments{
\item{path_to_unzip}{Path to unzip the book files}
}
\value{
TRUE, if sucessful
}
\description{
Donwload and unzip book file
}
\examples{
\dontrun{
flag <- afedR_unzip_book_file()
}
}
