% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_webex_fcts.R
\name{build_answers_text}
\alias{build_answers_text}
\title{Builds answer text}
\usage{
build_answers_text(text1, text2, text3)
}
\arguments{
\item{text1}{string}

\item{text2}{string}

\item{text3}{string}
}
\value{
a list
}
\description{
Builds answer text
}
