% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_fcts_pdf.R
\name{compile_pdf_exercises}
\alias{compile_pdf_exercises}
\title{Builds pdf exercices}
\usage{
compile_pdf_exercises(
  students_names,
  students_ids,
  class_name = "Sample Class",
  exercise_name = "Sample Exercise",
  chapters_to_include = 1:3,
  dir_out = "afedR-pdf-exercises",
  solution = FALSE
)
}
\arguments{
\item{students_names}{Names of students (a vector)}

\item{students_ids}{Ids of students (a vector)}

\item{class_name}{The name of the class}

\item{exercise_name}{The name of the exercises}

\item{chapters_to_include}{Chapter to include in exercise (1-13)}

\item{dir_out}{Folder to copy exercise html files}

\item{solution}{Wheter to include solution or not..}
}
\value{
a dtaframe with solutions
}
\description{
Builds pdf exercices
}
