% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R
\name{format_cash}
\alias{format_cash}
\title{Fromats cash values}
\usage{
format_cash(x, type_cash = "USD")
}
\arguments{
\item{x}{A numeric}

\item{type_cash}{Type of cash}
}
\value{
A formatted string
}
\description{
Fromats cash values
}
